requirements.txt
# Flask and web framework dependencies
Flask>=2.3.0,<3.1.0
Werkzeug>=2.3.0,<3.0.0

# Machine Learning dependencies
scikit-learn>=1.2.0,<1.4.0
pandas>=1.5.0,<3.0.0
numpy>=1.21.0,<2.0.0
imbalanced-learn>=0.10.0,<0.12.0

# Text processing
nltk>=3.7,<4.0

# Development and testing
pytest>=7.0.0,<8.0.0
pytest-flask>=1.2.0,<2.0.0

# Production deployment
gunicorn>=20.0.0,<22.0.0

# Monitoring and logging
psutil>=5.8.0,<6.0.0

# Utilities
python-dotenv>=0.19.0,<2.0.0
click>=8.0.0,<9.0.0